[badger]{attempt to integrate badger storage}

+++ internal/services/donationalerts/response/token.go
@@ -4,5 +4,5 @@ type Token struct {
 	AccessToken  string `json:"access_token"`
 	RefreshToken string `json:"refresh_token"`
 	TokenType    string `json:"token_type"`
-	ExpiresIn    int    `json:"expires_in"`
+	ExpiresIn    int64  `json:"expires_in"`
 }
